IPL 2008-24 Stats
Just wrapped up a super fun data viz project on IPL stats! 🏏🔍 From the highest wicket takers to the top run scorers, I uncovered some wild insights.

🔥 Mumbai Indians turned out to be the team with the most wickets taken, but also—plot twist—they’ve conceded the most runs through no-balls and leg byes!
🎯 Yuzvendra Chahal came out on top as the highest wicket-taker in IPL history.
🤯 And get this: the most common way batters get out? Caught out—by a mile!

It was a rollercoaster putting this together, and even though it took me around 3 hours (and a bit of redoing here and there), it was totally worth it. 📊💡